chore: split explorer tests
